[P1-5] Add new identity fields to ZhtpIdentity struct (#8)
* Adds comprehensive test suite for ZhtpIdentity fields

Validates new identity fields including DID, device mappings, cryptographic secrets, DAO membership, and citizenship verification

Tests ensure proper field types, serialization behavior, and multi-device support functionality

Covers acceptance criteria for enhanced identity structure with fixed-size arrays for secrets and HashMap for device node management

Relates to P1-5

* Implements deterministic identity field derivation

Refactors identity creation to derive all cryptographic fields deterministically from master keypair according to architecture specification

Adds proper derivation functions for ZK secrets, wallet seeds, and DAO member IDs using Blake3 hashing
Updates constructors to require private key for secure field generation
Enhances security by ensuring secrets are never serialized and derived consistently

Relates to #7

* Extends identity creation with additional fields

Updates identity and wallet creation methods to support enhanced user data including names, contact information, and cryptographic keys

Modifies wallet creation to use seed phrase generation methods and return both wallet and seed data

Enhances test identity creation with private keys, device information, and location data to support more comprehensive identity verification

Relates to #7

* Enhances identity security with proper secret derivation

Prevents sensitive cryptographic material from being serialized by marking secrets as skip-only fields that require explicit re-derivation after deserialization.

Adds citizenship verification parameter to constructor enabling proper DAO voting power assignment (10 for verified citizens, 1 for unverified humans).

Introduces rederive_secrets method to safely restore cryptographic secrets from private keys after loading identities from storage.

Updates test helpers to use proper constructor ensuring all cryptographic fields are correctly derived rather than hardcoded.

* Adds cryptographic secret validation for identity security

Introduces validation methods to verify that cryptographic secrets are properly derived after deserialization, preventing use of zero-valued secrets that could compromise security.

Adds comprehensive test coverage for deterministic key derivation, deserialization security patterns, and DAO voting power rules based on citizenship status.

Updates test infrastructure to use realistic Dilithium2 key sizes for more accurate cryptographic testing.

* Improves identity deserialization security model

Removes default serde attributes from non-sensitive fields to prevent silent zero initialization

Consolidates security-critical secret field handling by moving default functions and clarifying that secrets are always zero after deserialization

Updates test infrastructure to use realistic Dilithium2 cryptographic key sizes instead of placeholder values

Enhances documentation to emphasize the critical requirement for secret re-derivation after deserialization

* Enhances identity deserialization security and safety

Addresses critical security vulnerability where deserialized identities contain zero-valued cryptographic secrets that could compromise system security if used directly.

Adds safe deserialization helper method that automatically enforces secret re-derivation, preventing accidental use of compromised identities.

Expands documentation with clear security warnings, usage examples for both safe and unsafe patterns, and explicit guidance on proper identity construction.

Enhances test coverage with comprehensive golden vector validation to ensure deterministic secret derivation and proper security boundary enforcement.

* Prevents unsafe identity deserialization

Blocks direct deserialization of ZhtpIdentity struct to enforce secure secret re-derivation.

Removes auto-Deserialize trait implementation and adds custom deserializer that returns an error, forcing developers to use the secure from_serialized() method which properly re-derives cryptographic secrets.

Updates from_serialized() to manually parse JSON fields instead of relying on blocked deserializer, ensuring all derived fields are computed correctly through the new() constructor.

Enhances test coverage to verify direct deserialization fails with appropriate error message while safe deserialization continues working properly.

* feat(P1-5 + P1-6): Add identity fields with enforced derivation and align with Issue #9 spec

This commit addresses both P1-5 (Add new identity fields) and P1-6 (Cryptographic derivation functions).

## P1-5 Changes

### Security Enhancements
- Add custom Deserialize implementation blocking direct deserialization
- Add safe from_serialized() method with comprehensive field restoration
- Fix data loss bug in from_serialized() to restore all non-derived fields
- Fix dao_voting_power bug in create_citizen_identity() (now sets to 10 for verified citizens)

### Test Improvements
- Add golden vector tests with real Dilithium2 sizes (PK=1312, SK=2528 bytes)
- Use blake3 crate directly for independent validation of expected outputs
- Test three vectors: all-zeros, pattern (0xAB/0xEF), and determinism validation

### Results
- All 12 acceptance criteria tests passing
- All 9 integration tests passing

## P1-6 Changes (Issue #9 Alignment)

### Derivation Function Updates
1. **generate_did()**: Now uses public_key.key_id directly instead of blake3(dilithium_pk)
   - Per Issue #9 spec: did:zhtp:{hex(key_id)}
   - Updated callers: new(), from_legacy_fields()

2. **derive_credential_hash()**: Now requires age and jurisdiction (no optional defaults)
   - Changed signature: fn derive_credential_hash(secret: &[u8; 32], age: u64, jurisdiction: &str)
   - Updated callers to provide default values (25, US) when None
   - Updated all call sites: new(), from_legacy_fields(), rederive_secrets()

### Test Updates
- Updated golden vector test expected DID to use key_id format
- All tests passing with new derivation logic

## Files Modified
- src/identity/lib_identity.rs - Core identity implementation
- tests/zhtp_identity_fields.rs - Acceptance criteria tests
- examples/dao_wallet_demo.rs - Fix unrelated compilation error

+++ src/types/proof_params.rs
@@ -15,7 +15,7 @@ pub struct IdentityProofParams {
     /// Privacy level (0 = public, 100 = maximum privacy)
     pub privacy_level: u8,
     /// Minimum reputation score required
-    pub min_reputation: Option<u32>,
+    pub min_reputation: Option<u64>,
     /// Proof type identifier
     pub proof_type: String,
     /// Require citizenship status
@@ -45,7 +45,7 @@ impl IdentityProofParams {
     }
     
     /// Set minimum reputation requirement
-    pub fn with_min_reputation(mut self, min_reputation: u32) -> Self {
+    pub fn with_min_reputation(mut self, min_reputation: u64) -> Self {
         self.min_reputation = Some(min_reputation);
         self
     }
